@@ -12,42 +12,42 @@
     "coverage": "vitest run --coverage"
   },
   "dependencies": {
-    "@amplitude/analytics-browser": "^2.9.3",
+    "@amplitude/analytics-browser": "^2.11.9",
     "@amplitude/plugin-user-agent-enrichment-browser": "^1.0.1",
-    "@fluentui/react": "^8.119.0",
-    "@fluentui/react-charting": "^5.21.22",
-    "@navikt/ds-css": "^ 6.12.0",
-    "@navikt/ds-react": "^ 6.12.0",
+    "@fluentui/react": "^8.121.13",
+    "@fluentui/react-charting": "^5.23.17",
+    "@navikt/ds-css": "^ 7.5.3",
+    "@navikt/ds-react": "^ 7.5.3",
     "d3": "^7.9.0",
     "react": "^18.3.1",
     "react-circular-progressbar": "^2.1.0",
     "react-dom": "^18.3.1",
-    "react-router-dom": "^6.24.0"
+    "react-router-dom": "^7.0.1"
   },
   "devDependencies": {
-    "@testing-library/jest-dom": "^6.4.6",
-    "@testing-library/react": "^16.0.0",
-    "@types/react": "^18.3.3",
-    "@types/react-dom": "^18.3.0",
-    "@typescript-eslint/eslint-plugin": "^7.14.1",
-    "@typescript-eslint/parser": "^7.14.1",
-    "@vitejs/plugin-react-swc": "^3.7.0",
-    "@vitest/coverage-v8": "^1.6.0",
-    "@vitest/ui": "^1.6.0",
-    "autoprefixer": "^10.4.19",
+    "@testing-library/jest-dom": "^6.6.3",
+    "@testing-library/react": "^16.0.1",
+    "@types/react": "^18.3.12",
+    "@types/react-dom": "^18.3.1",
+    "@typescript-eslint/eslint-plugin": "^8.15.0",
+    "@typescript-eslint/parser": "^8.15.0",
+    "@vitejs/plugin-react-swc": "^3.7.1",
+    "@vitest/coverage-v8": "^2.1.5",
+    "@vitest/ui": "^2.1.5",
+    "autoprefixer": "^10.4.20",
     "dotenv": "^16.4.5",
-    "eslint": "^8.56.0",
-    "eslint-plugin-react-hooks": "^4.6.2",
-    "eslint-plugin-react-refresh": "^0.4.7",
-    "express": "^4.19.2",
+    "eslint": "^9.15.0",
+    "eslint-plugin-react-hooks": "^5.0.0",
+    "eslint-plugin-react-refresh": "^0.4.14",
+    "express": "4.21.1",
     "jest-image-snapshot": "^6.4.0",
-    "jsdom": "^24.1.0",
+    "jsdom": "^25.0.1",
     "path": "^0.12.7",
-    "postcss": "^8.4.38",
+    "postcss": "^8.4.49",
     "postcss-import": "^16.1.0",
-    "tailwindcss": "^3.4.4",
-    "typescript": "^5.5.2",
-    "vite": "^5.3.1",
-    "vitest": "^1.6.0"
+    "tailwindcss": "^3.4.15",
+    "typescript": "^5.7.2",
+    "vite": "5.4.11",
+    "vitest": "^2.1.5"
   }
 }
